JDK-8263322 Seperated the unit-tests into seperate files
