+++ tests/system/src/test/java/test/com/sun/javafx/application/InitializeJavaFXStartupBase.java
@@ -1,0 +1,17 @@
+package test.com.sun.javafx.application;
+
+import javafx.application.Platform;
+
+import java.util.concurrent.CountDownLatch;
+import java.util.concurrent.TimeUnit;
+
+public class InitializeJavaFXStartupBase extends InitializeJavaFXBase {
+
+    public static void initializeStartup() throws Exception {
+        CountDownLatch latch = new CountDownLatch(1);
+        Platform.startup(() -> {
+            latch.countDown();
+        });
+        latch.await(5, TimeUnit.SECONDS);
+    }
+}
